Split completion work into multiple files
