+++ crates/ark/src/lsp/backend.rs
@@ -29,13 +29,13 @@ use tower_lsp::LanguageServer;
 use tower_lsp::LspService;
 use tower_lsp::Server;
 
-use crate::lsp::completions::append_document_completions;
-use crate::lsp::completions::append_session_completions;
-use crate::lsp::completions::append_workspace_completions;
-use crate::lsp::completions::can_provide_completions;
-use crate::lsp::completions::completion_context;
-use crate::lsp::completions::resolve_completion_item;
-use crate::lsp::completions::CompletionData;
+use crate::lsp::completions::activate::can_provide_completions;
+use crate::lsp::completions::document::append_document_completions;
+use crate::lsp::completions::resolve::resolve_completion_item;
+use crate::lsp::completions::session::append_session_completions;
+use crate::lsp::completions::types::completion_context;
+use crate::lsp::completions::types::CompletionData;
+use crate::lsp::completions::workspace::append_workspace_completions;
 use crate::lsp::definitions::goto_definition;
 use crate::lsp::diagnostics;
 use crate::lsp::documents::Document;
