--- crates/ark/src/lsp/completions/completions-pipelines.rs
@@ -1,5 +1,0 @@
-
-fn is_pipe_operator(node: &Node) -> bool {
-    matches!(node.kind(), "%>%" | "|>")
-}
-
