+++ crates/ark/src/lsp/completions/mod.rs
@@ -1,0 +1,14 @@
+//
+// mod.rs
+//
+// Copyright (C) 2023 Posit Software, PBC. All rights reserved.
+//
+//
+
+pub mod activate;
+pub mod completion_item;
+pub mod document;
+pub mod resolve;
+pub mod session;
+pub mod types;
+pub mod workspace;
